Moved IpAddressUtils to Net and fixed the service locator call for it.
