+++ module/VuFind/src/VuFind/Net/IpAddressUtils.php
@@ -25,7 +25,7 @@
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     http://www.vufind.org  Main Page
  */
-namespace VuFind\IpAddress;
+namespace VuFind\Net;
 
 /**
  * IP address utility functions.
